+++ models/mobilenet-simplified-for-inference-shallow.relay
