+++ models/mobilenet-simplified-for-inference.relay
