+++ models/resnet-simplified-for-inference-shallow.relay
