+++ models/resnet-simplified-for-inference.relay
