+++ tests/mobilenet-relay-to-glenside.rs
@@ -19,7 +19,7 @@ use std::path::PathBuf;
 #[test]
 fn parse_mobilenet_simplified_for_inference() {
     let filename = PathBuf::from(format!(
-        "{}/models/mobilenet.relay",
+        "{}/models/mobilenet-simplified-for-inference.relay",
         env!("CARGO_MANIFEST_DIR")
     ));
     let relay = std::fs::read_to_string(&filename).unwrap();
